Add support for fetching and parsing Gradle module files. (#26)
Gradle publishes additional metadata in a file suffixed by .module, which supplants the maven pom during gradle builds. This is also the mechanism by which certain other features, such as kotlin multiplatform artifact referencing is done. This commit adds support for parsing those files, with some buffering against evolution of the data file format (versioned data classes, and a place to add selection by formatVersion once incompatible versions of the spec come out.  For the moment, the 1.1 spec is a strict superset, so 1.0 files are parseable by the same parser, into the same data objects. The model objects are kotlin data classes for the implementations, easily targeted by moshi for the parse outcomes. The parser is a moshi kotlin configuration.

No semantics are assumed in this model - no further resolution, or following available-at references. That is the responsibility of clients, to interpret according to gradle's specified semantics.