Strip out fine grained packaging. (#30)
The release pom/artifact builder stuff really assumes a single bazel-built .jar, in its current form, and was ill-designed to handle a bazel-style fine-grained dep system. While a good maven-artifact preparer system would be good, it's pretty out of scope, and this is supposed to be a small artifact anyway, self-contained except for its dependencies on maven model stuff itself. So just kill the fine-grained pom stuff, and point the tests at the main jar.